noaa_gsom_fetcher/requirements.txt
requests
python-dateutil
tqdm
influxdb
numpy